requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
pydantic
pulp
python-dateutil
statsmodels
